Finished CUbe in and Cube out and resolved requires problem

+++ src/Commands/CubeIn.h
@@ -8,7 +8,7 @@
 
 class CubeIn : public frc::Command {
 private:
-	static constexpr float FLYWHEEL_SPEED = 1;
+	static constexpr int FLYWHEEL_SPEED = 1;
 public:
 	CubeIn();
 	void Initialize();
